requirements.txt
# Django 核心
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5.0
djangorestframework-simplejwt>=5.3.0

# WebSocket & Channels
channels>=4.0.0
channels-redis>=4.1.0

# 缓存
django-redis>=5.4.0
redis>=5.0.0

# AI 识别 - YOLOv8
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0

# 数据库
psycopg2-binary>=2.9.0

# 生产环境
gunicorn>=21.0.0
uvicorn>=0.24.0
whitenoise>=6.0.0
dj-database-url>=2.0.0

# 工具
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pyngrok>=7.0.0
